Untouched program from 2014
This code is a horrible mess. But it technically works! Requires Python 3 with colorama, colormath and other modules.
I wrote this a long time ago, long before I had actually taken a programming course.
The only change I made to the file is a docstring at the top with a copyright notice.